+++ src/lib.rs
@@ -2,6 +2,7 @@
 pub mod config;
 pub mod database;
 pub(crate) mod extraction;
+pub mod manifest;
 pub(crate) mod parameters;
 pub(crate) mod samples;
 pub mod specs;