+++ src/specs/mod.rs
@@ -1,5 +1,6 @@
 #![doc = include_str!("../../GETTING_STARTED.md")]
 pub mod address_appearance_index;
+#[allow(dead_code)]
 pub(crate) mod my_data_spec;
 pub mod nametags;
 pub mod traits;